requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
Babel==2.6.0
baidu-aip==2.2.5.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.3
defusedxml==0.5.0
Flask==0.12.4
Flask-AppBuilder==1.11.1
Flask-Babel==0.11.1
Flask-Login==0.2.11
Flask-Migrate==2.2.1
Flask-OpenID==1.2.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.13
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
python3-openid==3.1.0
pytz==2018.5
pywin32==223
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
